Implement new format strings for ints: #1

+++ include/scn/detail/scan.h
@@ -319,7 +319,7 @@ namespace scn {
     {
         SCN_EXPECT(!str.empty());
         auto s = detail::integer_scanner<T>{};
-        s.base = base;
+        s.base = static_cast<uint8_t>(base);
         bool minus_sign = false;
         auto sp = make_span(str.data(), str.size()).as_const();
         if (str[0] == detail::ascii_widen<CharT>('-')) {
